Finished the project
Now waiting for the WiFi to connect and the micro:bit to act connected, if it's connected before the Raspberry Pi is booted it just treats it like a charging device, while I need it to treat it like a filesystem and serial host. Forcing user to un-/replug the micro:bit to work.